1130 Progress

+++ src/main/front-end/app_board/src/Article/FileForm.js
@@ -8,10 +8,15 @@ function Files({files, createdId}) {
 
     function deleteFile(id, filename){
         if(window.confirm(filename + " 을 삭제하시겠습니까? ")){
-            axios.delete(`/file/delete/${id}`).catch((e) => {
-                console.log(e.response.status + " : " + e.response.statusText);
-            })
-            window.location.reload();
+            const test = prompt("\"test\" 를 입력하세요");
+            if(_.isEqual(test, "test")){
+                axios.delete(`/file/delete/${id}`).catch((e) => {
+                    console.log(e.response.status + " : " + e.response.statusText);
+                })
+                window.location.reload();
+            }else{
+                alert("입력값이 올바르지 않습니다.\n삭제에 실패했습니다.");
+            }
         }
     }
